Add hooks and utility functions for page selection in settings editor
- Introduced `usePageSearch` and `useSelectedItem` hooks to manage page selection and search functionality.
- Added utility functions for formatting page items, enhancing the user experience in selecting pages.
- Created `Suffix` component to display loading states and allow removal of selected pages.
- Updated `SingleSelectPageWithSearchEdit` component to utilize new hooks and components for improved functionality.
