+++ packages/js/settings-editor/src/form-controls/single-select-page-with-search/types.ts
@@ -1,0 +1,7 @@
+/**
+ * The type for the page item for the select control.
+ */
+export type PageItem = {
+	label: string;
+	value: string;
+};
